refactor logic so we read previously-cleaned files if they exist so expectations match.

+++ test/datasets/test_iwslt2016.py
@@ -83,18 +83,32 @@ def _get_mock_dataset(root_dir, split, src, tgt, valid_set, test_set):
 
     mocked_data = defaultdict(lambda: defaultdict(list))
 
-    _, uncleaned_file_names = _generate_iwslt_files_for_lang_and_split(16, src, tgt, valid_set, test_set)
-    src_file = uncleaned_file_names[src][split]
-    tgt_file = uncleaned_file_names[tgt][split]
+    cleaned_file_names, uncleaned_file_names = _generate_iwslt_files_for_lang_and_split(16, src, tgt, valid_set, test_set)
+    uncleaned_src_file = uncleaned_file_names[src][split]
+    uncleaned_tgt_file = uncleaned_file_names[tgt][split]
 
-    for file_name in (src_file, tgt_file):
-        out_file = os.path.join(inner_temp_dataset_dir, file_name)
-        with open(out_file, "w") as f:
-            # Get file extension (i.e., the language) without the . prefix (.en -> en)
-            lang = os.path.splitext(file_name)[1][1:]
-            mocked_data_for_split, file_contents = _generate_uncleaned_contents(split)
-            mocked_data[split][lang] = mocked_data_for_split
-            f.write(file_contents)
+    cleaned_src_file = cleaned_file_names[src][split]
+    cleaned_tgt_file = cleaned_file_names[tgt][split]
+
+    for (unclean_file_name, clean_file_name) in [
+        (uncleaned_src_file, cleaned_src_file),
+        (uncleaned_tgt_file, cleaned_tgt_file)
+    ]:
+        # Get file extension (i.e., the language) without the . prefix (.en -> en)
+        lang = os.path.splitext(unclean_file_name)[1][1:]
+        expected_clean_filename = os.path.join(inner_temp_dataset_dir, clean_file_name)
+
+        # If we've already written a clean file, read it, so we don't generate
+        # new random strings. Otherwise generate new files and clean when read.
+        if os.path.exists(expected_clean_filename):
+            with open(expected_clean_filename, encoding="utf-8") as f:
+                mocked_data[(split, valid_set, test_set)][lang] = f.readlines()
+        else:
+            out_file = os.path.join(inner_temp_dataset_dir, unclean_file_name)
+            with open(out_file, "w") as f:
+                mocked_data_for_split, file_contents = _generate_uncleaned_contents(split)
+                mocked_data[(split, valid_set, test_set)][lang] = mocked_data_for_split
+                f.write(file_contents)
 
     inner_compressed_dataset_path = os.path.join(
         outer_temp_dataset_dir, f"{src}-{tgt}.tgz"
@@ -105,12 +119,12 @@ def _get_mock_dataset(root_dir, split, src, tgt, valid_set, test_set):
         tar.add(inner_temp_dataset_dir, arcname=f"{src}-{tgt}")
 
     outer_temp_dataset_path = os.path.join(
-        root_dir, "2016-01.tgz"
+        root_dir, "IWSLT2016", "2016-01.tgz"
     )
     with tarfile.open(outer_temp_dataset_path, "w:gz") as tar:
         tar.add(outer_temp_dataset_dir, arcname="2016-01")
 
-    return list(zip(mocked_data[split][src], mocked_data[split][tgt]))
+    return list(zip(mocked_data[(split, valid_set, test_set)][src], mocked_data[(split, valid_set, test_set)][tgt]))
 
 
 class TestIWSLT2016(TempDirMixin, TorchtextTestCase):
